backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==4.6.0
aioredis==2.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
openai==1.3.7
anthropic==0.7.8
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
python-dotenv==1.0.0
email-validator==2.1.0
httpx==0.25.2
aiofiles==23.2.1
pgvector==0.2.4
requests==2.31.0
transformers==4.35.0
torch==2.1.0
bitsandbytes==0.41.0
accelerate==0.24.0
# Cloud Provider Libraries
boto3==1.34.0
azure-monitor-query==1.2.0
google-cloud-logging==3.8.0
# Encryption
cryptography==41.0.7
# WebSocket support
websockets==12.0
aiosqlite==0.19.0
# Celery Task Queue
celery==5.3.4
celery[redis]==5.3.4
flower==2.0.1
# Additional integrations
sendgrid==6.11.0
jira==3.6.0
python-dateutil==2.8.2